+++ CHANGELOG.md
@@ -7,7 +7,10 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/) and this p
 ### Breaking
 ### Changed
 - Don't add event listeners if cluster module is not used.
+- Fixed issue with counters having extra records when using empty labels
 ### Added
+- Added `reset` to Counter and Gauge
+- Added `resetMetrics` to register to calling `reset` of all metric instances
 
 ## [10.1.1] - 2017-09-26
 ### Changed